feat(mcp): package MCP server as zettelforge.mcp module
Move the stdio MCP server from web/mcp_server.py into a proper
zettelforge.mcp package so `python -m zettelforge.mcp` works out
of a plain pip install (no git clone required).

- Add src/zettelforge/mcp/{__init__,__main__,server}.py
- Replace web/mcp_server.py with a thin back-compat shim so existing
  .mcp.json configurations keep working
- Wire `zettelforge` and `zettelforge-mcp` console scripts in
  pyproject.toml
- Point Homepage→docs.threatrecall.ai and add Changelog URL

All 7 MCP tools (remember, recall, synthesize, entity, graph, stats,
sync) are preserved bit-for-bit from the previous implementation.
